moved monitoring functions to legend-data-monitor
