+++ src/legend_data_monitor/__init__.py
@@ -3,6 +3,10 @@ from legend_data_monitor.analysis_data import AnalysisData
 from legend_data_monitor.core import control_plots
 from legend_data_monitor.slow_control import SlowControl
 from legend_data_monitor.subsystem import Subsystem
+from legend_data_monitor import monitoring 
+from legend_data_monitor import utils
+from legend_data_monitor import plot_styles
+from legend_data_monitor import plotting
 
 __all__ = [
     "__version__",
@@ -11,4 +15,8 @@ __all__ = [
     "AnalysisData",
     "SlowControl",
     "apply_cut",
+    "monitoring",  
+    "utils",       
+    "plot_styles",       
+    "plotting",       
 ]
