+++ src/legend_data_monitor/settings/final-calibrations.yaml
@@ -1,0 +1,9 @@
+{
+    "p04": ["r004"],
+    "p07": ["r008"], 
+    "p08": ["r015"],
+    "p09": ["r006"],
+    "p10": ["r007"],
+    "p11": ["r005"],
+    "p14": ["r006"]
+}
